+++ tests/testthat/test_compare.R
@@ -285,11 +285,8 @@ test_that("helper functions are working correctly", {
 
 test_that("Summary output looks right (i.e. for factors)", {
   expect_identical(
-    capture.output(summary(compare(mockstudy, mockstudy2, by = "case"))),
-    c(
-      ""                                                                              ,
-      ""                                                                              ,
-      "Table: Variables not shared"                                                   ,
+    capture.kable(summary(compare(mockstudy, mockstudy2, by = "case"))),
+    c("Table: Variables not shared"                                                   ,
       ""                                                                              ,
       "version   variable    position  class     "                                    ,
       "--------  ---------  ---------  ----------"                                    ,
@@ -374,11 +371,83 @@ test_that("Summary output looks right (i.e. for factors)", {
       "race    race    class  "                                                       ,
       "race    race    label  "                                                       ,
       "race    race    levels "                                                       ,
-      "bmi     bmi     label  "                                                       ,
-      ""
+      "bmi     bmi     label  "
     )
   )
 })
+
+
+test_that("Summary output with attributes and max.print options", {
+  expect_identical(
+    capture.kable(summary(compare(mockstudy, mockstudy2, by = "case"), show.attrs = TRUE,
+                          max.print.vars = 2, max.print.obs = 3, max.print.diff = 3, max.print.attrs = 3)),
+    c("Table: Variables not shared (5 differences not shown)"                        ,
+      ""                                                                             ,
+      "version   variable    position  class     "                                   ,
+      "--------  ---------  ---------  ----------"                                   ,
+      "x         age                2  integer   "                                   ,
+      "x         arm                3  character "                                   ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Other variables not compared"                                          ,
+      ""                                                                             ,
+      "var.x    pos.x  class.x     var.y    pos.y  class.y "                         ,
+      "------  ------  ----------  ------  ------  --------"                         ,
+      "race         5  character   race         3  factor  "                         ,
+      "ast         12  integer     ast          8  numeric "                         ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Observations not shared (1 differences not shown)"                     ,
+      ""                                                                             ,
+      "version     case   observation"                                               ,
+      "--------  ------  ------------"                                               ,
+      "x          88989             9"                                               ,
+      "x          90158             8"                                               ,
+      "x          99508             7"                                               ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Differences detected by variable"                                      ,
+      ""                                                                             ,
+      "var.x         var.y             n   NAs"                                      ,
+      "------------  ------------  -----  ----"                                      ,
+      "sex           sex            1495     0"                                      ,
+      "ps            ps                1     1"                                      ,
+      "hgb           hgb             266   266"                                      ,
+      "bmi           bmi               0     0"                                      ,
+      "alk.phos      alk.phos          0     0"                                      ,
+      "mdquality.s   mdquality.s       0     0"                                      ,
+      "age.ord       age.ord           0     0"                                      ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: First 3 differences detected per variable (1755 differences not shown)",
+      ""                                                                             ,
+      "var.x   var.y     case  values.x   values.y    row.x   row.y"                 ,
+      "------  ------  ------  ---------  ---------  ------  ------"                 ,
+      "sex     sex      76170  Male       Male           26      20"                 ,
+      "sex     sex      76240  Male       Male           27      21"                 ,
+      "sex     sex      76431  Female     Female         28      22"                 ,
+      "ps      ps       86205  0          NA              6       3"                 ,
+      "hgb     hgb      88714  NA         -9            192     186"                 ,
+      "hgb     hgb      88955  NA         -9            204     198"                 ,
+      "hgb     hgb      89549  NA         -9            229     223"                 ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Non-identical attributes (3 differences not shown)"                    ,
+      ""                                                                             ,
+      "var.x   var.y   name     attr.x                attr.y              "          ,
+      "------  ------  -------  --------------------  --------------------"          ,
+      "sex     sex     label    NA                    Sex (M/F)           "          ,
+      "sex     sex     levels   c(\"Male\", \"Female\")   c(\"Female\", \"Male\") "  ,
+      "race    race    class    NA                    factor              "
+    )
+  )
+})
+