+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * `compare()`: implement `n.diff.obs()`. (#124)
 
+* `summary.compare()`: allow for the display of attributes. (#125)
+
 * `tableby()`, `modelsum()`: properly propogated "term.name" to the `as.data.frame()` method. (#127, #128)
 
 * `tableby()`, `modelsum()`: Allow for LaTeX formatting. NOTE: this changes the formatting behavior when specifying